// server/src/utils/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Mutation = {
  __typename?: 'Mutation';
  createStage: Scalars['Boolean'];
  createTask: Scalars['Boolean'];
  deleteStage: Scalars['Boolean'];
  deleteTask: Scalars['Boolean'];
  editStage: Scalars['Boolean'];
  editTask: Scalars['Boolean'];
};
// ...
export type Query = {
  __typename?: 'Query';
  stage: Stage;
  stages: Array<Stage>;
  task: Task;
};


export type QueryStageArgs = {
  _id: Scalars['ID'];
};


export type QueryTaskArgs = {
  _id: Scalars['ID'];
};

export type Stage = {
  __typename?: 'Stage';
  _id: Scalars['ID'];
  createdAt: Scalars['String'];
  tasks?: Maybe<Array<Task>>;
  title: Scalars['String'];
  updatedAt?: Maybe<Scalars['String']>;
};

export type Task = {
  __typename?: 'Task';
  _id: Scalars['ID'];
  createdAt: Scalars['String'];
  isDone: Scalars['Boolean'];
  stageId: Scalars['String'];
  title: Scalars['String'];
  updatedAt?: Maybe<Scalars['String']>;
};
// ...
export type ResolversParentTypes = {
  Boolean: Scalars['Boolean'];
  ID: Scalars['ID'];
  Mutation: {};
  Query: {};
  Stage: Stage;
  String: Scalars['String'];
  Task: Task;
};